stupid bug in start up code because I still fail to remember that messages only go away after being ack'd, thus we should redeliver messages.
Also, reworked the stress gc test so that as before it really does ack messages in a non-linear order. This got quite a bit harder now we can't deliver arbitrary messages and need to build the mapping between msgid in the delivery and the seqid needed for the acks.

--HG--
branch : bug19662